add functionality to subheader
changed the way the views add their subheader content to the layout.
the layout view now has a 'subheaders' hash, which contains references
to each page's subheader content. this hash is populated in the
layout.initialize method, at the same time that the body of each
page is added to the layout. layout now has a method called
showSubheaderContent to select which view's subheader is to be shown.