Fixed a huge rename overrun. Enhanced server side order_by api and added order_by option to has_many.
Renamed _tuples on the client to storedTuples a while back. I missed that this also changed the
server code. Everything x_tuples got changed to xstoredTuples. Yuck.

Also added the ability to pass a "column_name direction" string in order_by on the server. Added
order_by as a has_many option. Added a spec for ordering.