Present notification of new election and new candidates in a single section
Added a new notification presenter that merges everything into a single
tree. This is in preparation to add comments to notifications.

New elections are shown with all their current answers, and elections that
have already been seen--that are just there to provide context for new
answers--are displayed in gray so as not to distract from new stuff.

Things are in a halfway state.