Fix bug with joining a selection on top of a join with subqueries.
Was generating qualified column names for the join, but predicates based
on the same columns were not using the qualified names. It was because
there was a selection above join which was associated with the subquery.
The join would resolve columns on the query even though it wasn't the
topmost operator corresponding to the query. Then the selection would
resolve the same columns without the qualified name, and these would get
used above it in the query instead of those created by the join. Now the
query decides when to generate a qualified select list, and it always
bases it on derived columns it pulls from the topmost relation in the
query. If that made any sense!